scsi/NCR5380: Handle BSY signal loss during information transfer phases
Improve robustness by adding checks for loss of the BSY signal during
the information transfer loop. The status register is being polled
anyway, so a BSY check costs nothing. BSY signal loss could be caused by
a target error or a kicked plug etc. A bus reset is another possibility
but that one is already handled, as hostdata->connected would be NULL.