scsi: NCR5380: Handle BSY signal loss during information transfer phases
Improve robustness by checking for a lost BSY signal during the
information transfer loop. The status register is being polled anyway,
so a BSY check costs nothing. BSY signal loss could be caused by a
target error or a kicked plug etc. A bus reset is another possibility
but that is already handled and hostdata->connected would be NULL.

Tested-by: Stan Johnson <userm57@yahoo.com>
Signed-off-by: Finn Thain <fthain@linux-m68k.org>