scsi/mac_scsi: Revise error handling
On my systems, failed PDMA transfers are infrequent but do occur and
need to be handled appropriately. When a PDMA transfer fails, the core
NCR5380 driver responds by setting the device 'borken' bit to inhibit
further DMA transfers. But that ruins thoughput, so better just report
the error to the mid-layer so the command will get retried. Moreover,
validating the Last Byte Sent flag makes that problem worse because the
flag is unreliable on this hardware.